% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrival_curve.R
\name{clean_and_join_nycflights}
\alias{clean_and_join_nycflights}
\title{Clean and join nycflights13 datasets}
\usage{
clean_and_join_nycflights(flights = NULL, planes = NULL, airports = NULL)
}
\arguments{
\item{flights}{A data frame containing flight data. If NULL, the function uses `nycflights13::flights`.}

\item{planes}{A data frame containing planes data. If NULL, the function uses `nycflights13::planes`.}

\item{airports}{A data frame containing airports data. If NULL, the function uses `nycflights13::airports`.}
}
\value{
A data frame with joined flights, planes, and airports data, including columns indicating if the airport is domestic US and flight type.
}
\description{
This function joins flights, planes, and airports data from the nycflights13 package.
It can also accept data frames as arguments for more flexibility.
}
