% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrival_curve.R
\name{clean_arrival_curve}
\alias{clean_arrival_curve}
\title{Clean Arrival Curve Data}
\usage{
clean_arrival_curve(arrival_curve)
}
\arguments{
\item{arrival_curve}{A data frame containing arrival curve data with a 'minutes_prior' column.}
}
\value{
A cleaned data frame with the following columns:
  \itemize{
    \item \code{minutes_prior}: Minutes prior to arrival time.
    \item Other columns representing aggregated and scaled values.
  }
}
\description{
This function cleans the arrival curve data by extracting numeric values from the 'minutes_prior' column,
grouping the data by minute intervals, summarizing the values, and scaling them to a percentage format.
}
\details{
This function extracts numeric values from the 'minutes_prior' column using regular expressions,
groups the data by minute intervals, sums the values within each interval, and scales the summed values
to a percentage format by dividing them by 100.
}
\examples{
arrival_curve <- data.frame(minutes_prior = c("10", "20", "30"),
                            value1 = c(5, 10, 15),
                            value2 = c(20, 30, 40))
clean_arrival_curve(arrival_curve)

}
