% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrival_curve.R
\name{create_arrival_curve_kde}
\alias{create_arrival_curve_kde}
\title{Create KDE Samples from Arrival Curve}
\usage{
create_arrival_curve_kde(arrival_curve_output)
}
\arguments{
\item{arrival_curve_output}{A data frame resulting from clean_arrival_curve(get_pgds_arrival_curve()).}
}
\value{
A data frame with KDE samples and additional metadata columns.
}
\description{
This function takes the output from clean_arrival_curve(get_pgds_arrival_curve()) and creates a KDE sample to join to the flight schedule.
}
