% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flightanalytics-package.R
\docType{package}
\name{flightanalytics-package}
\alias{flightanalytics}
\alias{flightanalytics-package}
\title{flightanalytics: Flight & Airport Analytics}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
